No Jira: moved certificate pages into certificate package
