+++ app/controllers/certificate/CertificateDeclarationSaoController.scala
@@ -20,7 +20,7 @@ import controllers.actions.*
 import forms.certificate.CertificateDeclarationSaoFormProvider
 import models.Mode
 import navigation.Navigator
-import pages.CertificateDeclarationSaoPage
+import pages.certificate.CertificateDeclarationSaoPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
