+++ app/controllers/certificate/CertificateReviewQualifiedController.scala
@@ -21,7 +21,7 @@ import controllers.routes
 import models.NormalMode
 import models.upload.*
 import navigation.Navigator
-import pages.{CertificateReviewQualifiedPage, CertificateSaoFullNamePage}
+import pages.certificate.{CertificateReviewQualifiedPage, CertificateSaoFullNamePage}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
