+++ app/controllers/certificate/CertificateReviewUnqualifiedController.scala
@@ -21,7 +21,7 @@ import controllers.routes
 import models.NormalMode
 import models.upload.*
 import navigation.Navigator
-import pages.{CertificateReviewUnqualifiedPage, CertificateSaoFullNamePage}
+import pages.certificate.{CertificateReviewUnqualifiedPage, CertificateSaoFullNamePage}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
