+++ app/controllers/certificate/CertificateSaoFullNameController.scala
@@ -20,7 +20,7 @@ import controllers.actions.*
 import forms.certificate.CertificateSaoFullNameFormProvider
 import models.Mode
 import navigation.Navigator
-import pages.CertificateSaoFullNamePage
+import pages.certificate.CertificateSaoFullNamePage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
