+++ app/controllers/certificate/CertificateWhoIsSubmittingController.scala
@@ -20,7 +20,7 @@ import controllers.actions.*
 import forms.certificate.CertificateWhoIsSubmittingFormProvider
 import models.Mode
 import navigation.Navigator
-import pages.CertificateWhoIsSubmittingPage
+import pages.certificate.CertificateWhoIsSubmittingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
