+++ app/controllers/certificate/QualifiedCompaniesController.scala
@@ -19,7 +19,7 @@ package controllers.certificate
 import controllers.actions.*
 import models.NormalMode
 import navigation.Navigator
-import pages.QualifiedCompaniesPage
+import pages.certificate.QualifiedCompaniesPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
