+++ app/controllers/certificate/UnqualifiedCompaniesController.scala
@@ -19,7 +19,7 @@ package controllers.certificate
 import controllers.actions.*
 import models.NormalMode
 import navigation.Navigator
-import pages.UnqualifiedCompaniesPage
+import pages.certificate.UnqualifiedCompaniesPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
