+++ app/pages/certificate/CertificateDeclarationSaoPage.scala
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package pages
+package pages.certificate
 
 import pages.Page.CERTIFICATE_PATH
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 
-case object CertificateSaoEmailPage extends QuestionPage[String] {
+case object CertificateDeclarationSaoPage extends QuestionPage[String] {
 
   override def path: JsPath = JsPath \ CERTIFICATE_PATH \ toString
 
-  override def toString: String = "certificateSaoEmail"
+  override def toString: String = "certificateDeclarationSao"
 }
