+++ app/pages/certificate/CertificateReviewQualifiedPage.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package pages
+package pages.certificate
 
 import pages.Page.CERTIFICATE_PATH
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 
-case object CertificateAdditionalInformationPage extends QuestionPage[Option[String]] {
-
+case object CertificateReviewQualifiedPage extends QuestionPage[Boolean] {
   override def path: JsPath = JsPath \ CERTIFICATE_PATH \ toString
 
-  override def toString: String = "certificateAdditionalInformation"
+  override def toString: String = "certificateReviewQualifiedPage"
 }
