+++ app/pages/certificate/CertificateReviewUnqualifiedPage.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package pages
+package pages.certificate
 
 import pages.Page.CERTIFICATE_PATH
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 
-case object CertificateSaoFullNamePage extends QuestionPage[String] {
-
+case object CertificateReviewUnqualifiedPage extends QuestionPage[Boolean] {
   override def path: JsPath = JsPath \ CERTIFICATE_PATH \ toString
 
-  override def toString: String = "certificateSaoFullName"
+  override def toString: String = "certificateReviewUnqualifiedPage"
 }
