+++ app/pages/certificate/CertificateSaoFullNamePage.scala
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package pages
+package pages.certificate
 
-import models.certificate.CertificateDeclarationStandIn
 import pages.Page.CERTIFICATE_PATH
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 
-case object CertificateDeclarationStandInPage extends QuestionPage[CertificateDeclarationStandIn] {
+case object CertificateSaoFullNamePage extends QuestionPage[String] {
 
   override def path: JsPath = JsPath \ CERTIFICATE_PATH \ toString
 
-  override def toString: String = "certificateDeclarationStandIn"
+  override def toString: String = "certificateSaoFullName"
 }
