+++ app/pages/certificate/CertificateWhoIsSubmittingPage.scala
@@ -14,14 +14,16 @@
  * limitations under the License.
  */
 
-package pages
+package pages.certificate
 
+import models.certificate.CertificateWhoIsSubmitting
 import pages.Page.CERTIFICATE_PATH
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 
-case object CertificateDeclarationSaoPage extends QuestionPage[String] {
+case object CertificateWhoIsSubmittingPage extends QuestionPage[CertificateWhoIsSubmitting] {
 
   override def path: JsPath = JsPath \ CERTIFICATE_PATH \ toString
 
-  override def toString: String = "certificateDeclarationSao"
+  override def toString: String = "certificateWhoIsSubmitting"
 }
