+++ app/viewmodels/checkAnswers/certificate/CertificateDeclarationSaoSummary.scala
@@ -18,7 +18,7 @@ package viewmodels.checkAnswers.certificate
 
 import controllers.certificate.routes as certificateRoutes
 import models.{CheckMode, UserAnswers}
-import pages.CertificateDeclarationSaoPage
+import pages.certificate.CertificateDeclarationSaoPage
 import play.api.i18n.Messages
 import uk.gov.hmrc.govukfrontend.views.viewmodels.summarylist.SummaryListRow
 import viewmodels.converters.*
