+++ it/test/repositories/SessionRepositoryISpec.scala
@@ -28,7 +28,7 @@ import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.must.Matchers
 import org.scalatestplus.mockito.MockitoSugar
 import org.slf4j.MDC
-import pages.CertificateUploadStatePage
+import pages.certificate.CertificateUploadStatePage
 import pages.notification.NotificationUploadStatePage
 import play.api.libs.json.Json
 import uk.gov.hmrc.mongo.test.DefaultPlayMongoRepositorySupport
