+++ test/controllers/certificate/CertificateAdditionalInformationControllerSpec.scala
@@ -26,7 +26,7 @@ import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.CertificateAdditionalInformationPage
+import pages.certificate.CertificateAdditionalInformationPage
 import play.api.data.Form
 import play.api.inject.bind
 import play.api.mvc.Call
