+++ test/controllers/certificate/CertificateSaoFullNameControllerSpec.scala
@@ -25,7 +25,7 @@ import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.CertificateSaoFullNamePage
+import pages.certificate.CertificateSaoFullNamePage
 import play.api.data.Form
 import play.api.inject.bind
 import play.api.mvc.Call
