+++ test/controllers/certificate/CertificateWhoIsSubmittingControllerSpec.scala
@@ -26,7 +26,7 @@ import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.CertificateWhoIsSubmittingPage
+import pages.certificate.CertificateWhoIsSubmittingPage
 import play.api.data.Form
 import play.api.inject.bind
 import play.api.mvc.Call
